requirements.txt
fastapi
uvicorn
python-multipart
chromadb
pdfplumber 
python-docx
requests
langchain-text-splitters
langchain-community
langchain
InstructorEmbedding 
sentence-transformers
langchain-huggingface
